+++ src/idvc/ui/windows.py
@@ -30,6 +30,13 @@ class VisualisationWindow(QtWidgets.QMainWindow):
         self.parent = parent
         self.setMinimumSize(200,200)
 
+    def createPopupMenu(self):
+        '''returns an empty menu for the main window to use as a popup menu.
+        
+        https://doc.qt.io/qt-6/qmainwindow.html#createPopupMenu
+        '''
+        return QtWidgets.QMenu(self)
+    
 class VisualisationWidget(QtWidgets.QMainWindow):
     '''creates a window with a QCILViewerWidget as the central widget
     '''
@@ -55,8 +62,7 @@ class VisualisationWidget(QtWidgets.QMainWindow):
 
     def getViewerType(self):
         return self.viewer
-
-        
+     
     def createEmptyFrame(self):
         #print("empty")
         self.frame = QCILViewerWidget(viewer=self.viewer, shape=(600,600), interactorStyle=self.interactorStyle)
@@ -227,6 +233,7 @@ class GraphsWindow(QMainWindow):
     
             GraphWidget = SingleRunResultsWidget(self, result, displ_wrt_point0)
             dock1 = QDockWidget(result.title,self)
+            dock1.setFeatures(QDockWidget.NoDockWidgetFeatures) 
             dock1.setAllowedAreas(QtCore.Qt.RightDockWidgetArea)
             dock1.setWidget(GraphWidget)
             self.addDockWidget(QtCore.Qt.RightDockWidgetArea,dock1)
@@ -243,10 +250,15 @@ class GraphsWindow(QMainWindow):
         
         SummaryTab = SummaryGraphsWidget(self, result_list)
         dock = QDockWidget("Summary",self)
+        dock.setFeatures(QDockWidget.NoDockWidgetFeatures) 
         dock.setAllowedAreas(QtCore.Qt.RightDockWidgetArea)
         dock.setWidget(SummaryTab)
         self.addDockWidget(QtCore.Qt.RightDockWidgetArea,dock)
         self.tabifyDockWidget(prev,dock)
 
         dock.raise_() # makes summary panel the one that is open by default.
+
+        # Stop the widgets in the tab to be moved around
+        for wdg in self.findChildren(QTabBar):
+            wdg.setMovable(False)